extraction/azure_function/core_function/requirements.txt
requests
beautifulsoup4
azure-storage-blob